Исправил округления

+++ internal/app/dto/get_balance_response.go
@@ -2,5 +2,5 @@ package dto
 
 type BalanceResponseBody struct {
 	Current   float64 `json:"current"`
-	Withdrawn int64   `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
